:recycle: chore: refactor styles properties references
