+++ src/pages/Dashboard/SideBar/styles.ts
@@ -16,7 +16,7 @@ export const AnimatedContainer = styled(motion.section)`
     max-height: 60rem;
     width: 100%;
     height: 100%;
-    border-radius: ${theme.radius.default};
+    border-radius: ${theme.radii.default};
     box-shadow: ${theme.shadows.default};
     padding: 2.4rem;
   `}
@@ -27,12 +27,12 @@ export const AnimatedAmountInvested = styled(motion.button)`
     ${({ theme }) => css`
       display: flex;
       justify-content: space-between;
-      color: ${theme.colors.lightGrey};
+      color: ${theme.colors.label};
       font-size: ${theme.fontSizes.small};
       width: 100%;
       margin: 0.8rem 0;
       padding: 0.8rem 2.4rem;
-      border-radius: ${theme.radius.small};
+      border-radius: ${theme.radii.small};
       transition: box-shadow 300ms ease-out;
       background: ${`linear-gradient(90deg, ${theme.colors.secondary} 0%, ${theme.colors.primary} 100%);`};
